+++ CHANGELOG.md
@@ -8,8 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Develop
 
 ### Added
-- Spike-in based scaling using DEseq2, spike-in counts, or Bonhoure et al.
-  approaches
+- Spike-in based scaling using DEseq2, spike-in counts, edgeR TMM, or Bonhoure
+  et al.  approaches
 - Deduplication with Picard MarkDuplicates.
 
 ### Changes